refactor: place list and organize code structure
