+++ app/src/main/java/ca/tanle/mapluv/ui/activities/viewmodels/CoordinatesViewModel.kt
@@ -1,4 +1,4 @@
-package ca.tanle.mapluv.ui.activities
+package ca.tanle.mapluv.ui.activities.viewmodels
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
