+++ app/src/main/java/ca/tanle/mapluv/ui/fragments/addPlace/AddressFragment.kt
@@ -1,27 +1,20 @@
-package ca.tanle.mapluv.ui.fragments
+package ca.tanle.mapluv.ui.fragments.addPlace
 
 import android.os.Bundle
-import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.core.widget.addTextChangedListener
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import ca.tanle.mapluv.R
-import ca.tanle.mapluv.data.remote.Address
-import ca.tanle.mapluv.data.remote.Geometry
-import ca.tanle.mapluv.data.remote.LocationR
 import ca.tanle.mapluv.databinding.FragmentAddressBinding
 import ca.tanle.mapluv.network.AddressRepository
 import ca.tanle.mapluv.network.RetrofitProvider
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+//import ca.tanle.mapluv.ui.fragments.ARG_PARAM1
+//import ca.tanle.mapluv.ui.fragments.ARG_PARAM2
+import ca.tanle.mapluv.ui.fragments.adapters.SearchAddressAdapter
 
 class AddressFragment : Fragment() {
     // TODO: Rename and change types of parameters
@@ -35,12 +28,9 @@ class AddressFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+//            param1 = it.getString(ARG_PARAM1)
+//            param2 = it.getString(ARG_PARAM2)
         }
-
-
-
     }
 
     override fun onCreateView(
@@ -87,8 +77,8 @@ class AddressFragment : Fragment() {
         fun newInstance(param1: String, param2: String) =
             AddressFragment().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+//                    putString(ARG_PARAM1, param1)
+//                    putString(ARG_PARAM2, param2)
                 }
             }
     }
