+++ app/src/main/java/ca/tanle/mapluv/ui/fragments/addPlace/CoordinatesFragment.kt
@@ -1,22 +1,17 @@
-package ca.tanle.mapluv.ui.fragments
+package ca.tanle.mapluv.ui.fragments.addPlace
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.get
-import ca.tanle.mapluv.R
 import ca.tanle.mapluv.databinding.FragmentCoordinatesBinding
 import ca.tanle.mapluv.ui.activities.AddPlaceActivity
-import ca.tanle.mapluv.ui.activities.CoordinatesViewModel
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+import ca.tanle.mapluv.ui.activities.viewmodels.CoordinatesViewModel
+//import ca.tanle.mapluv.ui.fragments.ARG_PARAM1
+//import ca.tanle.mapluv.ui.fragments.ARG_PARAM2
 
 class CoordinatesFragment : Fragment() {
     private var param1: String? = null
@@ -27,8 +22,8 @@ class CoordinatesFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+//            param1 = it.getString(ARG_PARAM1)
+//            param2 = it.getString(ARG_PARAM2)
         }
     }
 
@@ -76,8 +71,8 @@ class CoordinatesFragment : Fragment() {
         fun newInstance(param1: String, param2: String) =
             CoordinatesFragment().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+//                    putString(ARG_PARAM1, param1)
+//                    putString(ARG_PARAM2, param2)
                 }
             }
     }
