+++ app/src/main/res/layout/fragment_address.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragments.AddressFragment">
+    tools:context=".ui.fragments.addPlace.AddressFragment">
 
     <LinearLayout
         android:id="@+id/linearLayout4"
