+++ app/src/main/res/layout/fragment_coordinates.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.fragments.CoordinatesFragment">
+    tools:context=".ui.fragments.addPlace.CoordinatesFragment">
 
     <Button
         android:id="@+id/addPlaceBtn"
